+++ src/prime_rl/trainer/models/layers/lora/multi_linear.py
@@ -1,13 +1,9 @@
 import math
-from typing import TYPE_CHECKING
 
 import torch
 from torch import nn
 
 from .base import LoRAModule
-
-if TYPE_CHECKING:
-    from prime_rl.trainer.runs import Runs
 
 OFFSETS: torch.Tensor | None = None
 
@@ -125,24 +121,6 @@ class MultiLoRALinear(LoRAModule):
         else:
             nn.init.kaiming_uniform_(self.lora_A[index], a=math.sqrt(5))
             nn.init.zeros_(self.lora_B[index])
-
-    def register_with_runs(self, runs: "Runs", prefix: str) -> None:
-        """Register this module with the Runs system.
-
-        This method should be called after FSDP/compile/AC setup as these
-        transformations may change the underlying parameters while preserving
-        the module.
-
-        The Runs class will use this registration to:
-        - Get named parameters for specific run indices (for optimizer setup)
-        - Reset parameters when new runs are created
-        - Construct sliced state dicts for weight broadcast
-
-        Args:
-            runs: The Runs instance to register with
-            prefix: The module's name/prefix in the model (e.g., "model.layers.0.self_attn.q_proj")
-        """
-        runs.register_module(prefix, self)
 
     def forward(self, x: torch.Tensor) -> torch.Tensor:
         """
